fix edge sharing remove the "don't share edges if they have different contents" thing

+++ include/qbsp/qbsp.hh
@@ -327,7 +327,7 @@ class mapentity_t;
 struct face_fragment_t
 {
     winding_t w;
-    std::vector<size_t> edges; // only filled in MakeFaceEdges
+    std::vector<int64_t> edges; // only filled in MakeFaceEdges
     std::optional<size_t> outputnumber; // only valid for original faces after
                                         // write surfaces
 };
